Refactor the IO API
